+++ src/test/java/com/bsu/test_automation_framework_github/BasePageTest.java
@@ -1,0 +1,35 @@
+package com.bsu.test_automation_framework_github;
+
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+
+import java.util.concurrent.TimeUnit;
+
+public class BasePageTest {
+    private final String URL = "https://github.com/";
+    protected WebDriver driver;
+    protected MainPage mainPage;
+
+    @BeforeEach
+    public void setUp() {
+        WebDriverManager.chromedriver().setup();
+        driver = new ChromeDriver();
+
+        driver.manage().window().maximize();
+        driver.manage().timeouts().implicitlyWait(5000, TimeUnit.MILLISECONDS);
+        driver.manage().timeouts().pageLoadTimeout(10000, TimeUnit.MILLISECONDS);
+        driver.manage().timeouts().setScriptTimeout(5000, TimeUnit.MILLISECONDS);
+
+        driver.get(URL);
+
+        mainPage = new MainPage(driver);
+    }
+
+    @AfterEach
+    public void tearDown() {
+        driver.quit();
+    }
+}
